% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acc.R
\name{acc}
\alias{acc}
\title{acc}
\usage{
acc(old, new)
}
\arguments{
\item{old}{a numeric value}

\item{new}{a numeric value}
}
\value{
returns the an evaluation of old/new > U where U is a draw from the uniform distribution
}
\description{
Function to accept (or not) a proposed solution used as part of the MCMC procedure
}
\details{
A function for the evaluation of two likelihoods as part of the MCMC procedure
}
