% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.mod.R
\docType{data}
\name{bin.mod}
\alias{bin.mod}
\title{Example model for a survival outcome}
\format{
A model of class 'glm':
\describe{
 \item{vi}{vascular invasion}
 \item{ecog}{ECOG performance Status}
 \item{logafp}{AFP - log scale}
 \item{alb}{albumin}
 \item{logcreat}{Creatinine - log scale}
 \item{allmets}{metastesis}
}
}
\source{
simulated
}
\usage{
bin.mod
}
\description{
A generated model with a binary endpoint and a logistic link function.  Data
for the model were synthetically generated and are based on a dataset to
evaulate the use of Sorafenib in HCC akin to the PROSASH model
(see ?psc::surv.mod for more details)
}
\keyword{datasets}
