% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sest.R
\name{boot_sest}
\alias{boot_sest}
\title{Counter Factual Model - summary}
\usage{
boot_sest(
  i,
  pscOb = pscOb,
  lam = lam,
  kn = kn,
  k = k,
  cov = cov,
  tm = tm,
  rest = rest,
  beta = beta
)
}
\arguments{
\item{i}{indicator object}

\item{pscOb}{a pscOb object}

\item{lam}{parameters of the flexible spline model}

\item{kn}{knots included in the flexible spline model}

\item{k}{number of knots in the flexible spline model}

\item{cov}{a matrix of covariates}

\item{tm}{time at which to assess the survival function}

\item{rest}{a set of parameter covariate draws}

\item{beta}{parameter with which to adjust the baseline function}
}
\value{
A set of survival estimates
}
\description{
A function to estimate the survival function based on parameter estimates -
used in ootstrapping CFM for CIs
}
