% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmDataSumm.R
\name{cfmDataSumm}
\alias{cfmDataSumm}
\title{Summarising data within a CFM}
\usage{
cfmDataSumm(cfm)
}
\arguments{
\item{cfm}{a 'glm' or 'flexsurvreg' model object}
}
\value{
a summary table
}
\description{
The pscCFM creates a model object which is stripped of identifiable
information.  The cfmDataSumm fucntion supplies a tabulated form of the dataset
for summary information
}
