% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmDataVis.R
\name{cfmDataVis}
\alias{cfmDataVis}
\title{Visualising data within a CFM}
\usage{
cfmDataVis(cfm)
}
\arguments{
\item{cfm}{a 'glm' or 'flexsurvreg' model object}
}
\value{
a list of grobs for each model covariate
}
\description{
The pscCFM creates a model object which is stripped of identifiable
information.  The cfmDataVis function supplies a visualised form of the dataset
for summary information
}
