% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmDataVis_fac.R
\name{cfmDataVis_fac}
\alias{cfmDataVis_fac}
\title{Visualising Categorical Data}
\usage{
cfmDataVis_fac(x, nm)
}
\arguments{
\item{x}{a covariate to be summarised}

\item{nm}{a covariate name}
}
\value{
a ggplot object
}
\description{
A function which summarises categorical data using a bar plot.  A
sub-function of cfmDataVis
}
