% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmDataVis_num.R
\name{cfmDataVis_num}
\alias{cfmDataVis_num}
\title{Visualising Numerical Data}
\usage{
cfmDataVis_num(x, nm)
}
\arguments{
\item{x}{a covariate to be summarised}

\item{nm}{a covariate name}
}
\value{
a ggplot object
}
\description{
A function which summarises categorical data using density plots.  A
sub-function of cfmDataVis
}
