% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmSett.R
\name{cfmSett}
\alias{cfmSett}
\title{A function to store setting information for a cfm model}
\usage{
cfmSett(P = NULL, I = NULL, C = NULL, O = NULL)
}
\arguments{
\item{P}{Character; patient group}

\item{I}{Character; intervention group - perhaps left unspecified}

\item{C}{Character; comparitor group}

\item{O}{Character; outcome definition}
}
\description{
This function collects infomation ont he setting of a model for the
definition of a Counter Factual Model (CFM) object.  Setting information is
provided in the PICO format.  Intended to be used in conjuction with the
pscCRM.R function
}
\details{
A function for containg basic information pertaining to the setting
of a model in PICO format
}
