% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmSumm.flexsurvreg.R
\name{cfmSumm.flexsurvreg}
\alias{cfmSumm.flexsurvreg}
\title{Counter Factual Model - summary}
\usage{
cfmSumm.flexsurvreg(pscOb, bootCI = TRUE, nboot = 1000)
}
\arguments{
\item{pscOb}{an object of class 'psc'}

\item{bootCI}{a boolean to determine if bootstrapping CIs are required}

\item{nboot}{Number of bootstraps}
}
\value{
A summary of a cfm object
}
\description{
A generic function to provide a summary of a Counter factual model of class
'glm'
}
