% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmSumm.glm.R
\name{cfmSumm.glm}
\alias{cfmSumm.glm}
\title{Counter Factual Model - summary}
\usage{
cfmSumm.glm(pscOb, bootCI = TRUE, nboot = 1000, resp = TRUE)
}
\arguments{
\item{object}{an object of class 'psc'}
}
\value{
A summary of a cfm object
}
\description{
A generic function to provide a summary of a Counter factual model of class
'glm'
}
