% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmSumm.R
\name{cfmSumm}
\alias{cfmSumm}
\title{Counter Factual Model - summary}
\usage{
cfmSumm(pscOb)
}
\arguments{
\item{object}{an object of class 'psc'}
}
\value{
A summary of a psc object obtained using pscSumm and a copy of the pscfit object
}
\description{
A generic function to provide a summary of a Counter factual model
}
