% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmValid.R
\name{cfmValid}
\alias{cfmValid}
\title{Validation of a counter-factual model
This function intends to perform basic validation of a coutnerfactual model of
class 'glm' or 'flexsurvreg'.  As a default internal valudation will be
performed by estimating basic measures of discrimination and calibration.
If provided, measures will be estimated on external data. This is intended
to be used in conjunction with the pscCRM.R function#' A generic function for
 extracting model information}
\usage{
cfmValid(CFM)
}
\arguments{
\item{CFM}{a model object to be validated}
}
\description{
Validation of a counter-factual model
This function intends to perform basic validation of a coutnerfactual model of
class 'glm' or 'flexsurvreg'.  As a default internal valudation will be
performed by estimating basic measures of discrimination and calibration.
If provided, measures will be estimated on external data. This is intended
to be used in conjunction with the pscCRM.R function#' A generic function for
 extracting model information
}
\examples{
library(psc)
cfmValid(psc::surv.mod)
}
