% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.psc.R
\name{coef.psc}
\alias{coef.psc}
\title{Returns the coefficeint estimate of a psc object.}
\usage{
\method{coef}{psc}(x, ...)
}
\arguments{
\item{x}{a 'psc' object}
}
\value{
The summary of the posterior distribution for the efficacy parameter in terms of the median and 95% HPD
}
\description{
Returns the coefficeint estimate of a psc object.
}
