% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.mod.R
\docType{data}
\name{count.mod}
\alias{count.mod}
\title{Example model for a survival outcome}
\format{
A model of class 'glm':
\describe{
 \item{ecog}{ECOG performance Status}
 \item{logafp}{AFP - log scale}
 \item{alb}{albumin}
 \item{logcreat}{Creatinine - log scale}
}
}
\source{
simulated
}
\usage{
count.mod
}
\description{
A generated model with a count data endpoint and a log link function.
Data for the model were synthetically generated and are based on a dataset to
evaulate the use of Sorafenib in HCC akin to the PROSASH model
(see ?psc::surv.mod for more details)
}
\keyword{datasets}
