% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.comb.R
\name{data.comb.fpm}
\alias{data.comb.fpm}
\title{Functions for cleaning a dataset to match the output of a CFM}
\usage{
data.comb.fpm(model.extract, DC, id = NULL)
}
\arguments{
\item{model.extract}{details of model extract from model extract}

\item{DC}{A dataset to compare against the CFM}
}
\value{
A cleaned dataset and outcome
}
\description{
Functions for cleaning a dataset to match the output of a CFM
}
