% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{data_match}
\alias{data_match}
\title{A function to ensure that data from the cfm and data cohort are compatible}
\usage{
data_match(cls, lev, dc.data)
}
\arguments{
\item{cls}{a list of extracted data classes}

\item{lev}{a list of factor levels}

\item{dc.data}{dataset to be 'cleaned'}
}
\value{
a dataset which is checked and compatible with the CFM
}
\description{
The purpose of this function is to run a series of checks to ensure that the
data included in the data cohort is comparable to the counter-factual model
}
