% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataComb.flexsurvreg.R
\name{dataComb.flexsurvreg}
\alias{dataComb.flexsurvreg}
\title{Fucntion for cleaning the data of a model with class 'flexsurvreg'}
\usage{
\method{dataComb}{flexsurvreg}(CFM, DC, id = NULL, trt = NULL)
}
\arguments{
\item{CFM}{a model object supplied to pscfit}

\item{DC}{a dataset including covariates to match the CFM}

\item{id}{a vector specifiying whether a subset of the dataset should be selected.
Defaults to 'NULL' e.g. all data points included}

\item{trt}{An optional additional vector denoting treatment allocations for multiple treatment comparisons.  Defaults to 'NULL'}
}
\value{
a list containing objects which specifiy the required exported components
  of the model and a cleaned data cohort.
}
\description{
Fucntion for cleaning the data of a model with class 'flexsurvreg'
}
