% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataComb.glm.R
\name{dataComb.glm}
\alias{dataComb.glm}
\title{Fucntion for cleaning the data of a model with class 'flexsurvreg'}
\usage{
\method{dataComb}{glm}(CFM, DC, id = NULL)
}
\arguments{
\item{CFM}{a model object supplied to pscfit}

\item{DC}{a dataset including covariates to match the CFM}

\item{id}{a vector specifiying whether a subset of the dataset shoudl be selcted.
Defaults to 'NULL' e.g. all data points included}
}
\value{
a list containing objects which specifiy the required exported components
  of the model and a cleaned data cohort.
}
\description{
Fucntion for cleaning the data of a model with class 'flexsurvreg'
}
