% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataComb.pscCFM.R
\name{dataComb.pscCFM}
\alias{dataComb.pscCFM}
\title{Function for cleaning the data of a model with class}
\usage{
\method{dataComb}{pscCFM}(CFM, DC, id = NULL, trt = NULL, cfmOb = TRUE)
}
\arguments{
\item{CFM}{a model object supplied to pscfit}

\item{DC}{a dataset including covariates to match the CFM}

\item{id}{to specify which observations in the data cohort should be evaluated.
Defualts to 'NULL' i.e all observations}

\item{trt}{used to specify multiple treatment effects. Defaults to NULL}

\item{cfmOb}{used to specify if a CFM object is supplies as the counter factual model}
}
\value{
a list containing objects which specifiy the required exported components
  of the model and a cleaned data cohort.  Attirbutes of the 'cleaned' object include:
 Attributes include \itemize{
 \item{'model.type' specifying the class of model to be used as the CFM }
 \item{'model_extract' sepcifying the model componets required for estimation}
 \item{'cov' a cleaned dataset of covariates}
 \item{'outcome' a cleaned dataset containing the outcomes}
 }
}
\description{
The purpose of this function is to prepare the dataset and the counter-factual
model for estimation and is the first step pf the pscfit.R process. The output
is a complete-case dataset where the data names match the variables used in the CFM.
}
