% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataSumm.R
\name{dataSumm}
\alias{dataSumm}
\title{A function which provides basic summary information of a matrix x}
\usage{
dataSumm(x)
}
\arguments{
\item{x}{a matrix of covariates}
}
\description{
A function which provides basic summary information of a matrix x
}
\details{
Categorical/Character data are summarised by a table and Continuous
    data are summarised as median (IQR)
}
