% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e4_data.R
\docType{data}
\name{e4_data}
\alias{e4_data}
\title{Example Dataset of patients treated with GemCap in the ESPAC-4 trial}
\format{
A model of class 'flezsurvreg':
\describe{
 \item{time}{survival time}
 \item{cen}{censoring indicator}
 \item{nodes}{negative (n=1) or positive (n=2) lymph nodes}
 \item{grade}{tumour grade (1,2 or 3)}
 \item{lca199}{log transformed ca19.9}
 \item{t}{T-stage (1,2 or 3)}
}
}
\source{
simulated
}
\usage{
e4_data
}
\description{
A dataset containing 346 simulated patients.  Data are based on the patietns
randomised to revceive GemCap in the ESPAC-4 trial
}
\keyword{datasets}
