% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facVisComp.R
\name{facVisComp}
\alias{facVisComp}
\title{Visualising Categorical Data}
\usage{
facVisComp(p, x)
}
\arguments{
\item{p}{a ggplot objects}

\item{x}{a categorical covariate}
}
\value{
a ggplot object
}
\description{
A function which compares visually a new categorical covariate against
equivalent data from a CFM
}
