% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParm.R
\name{initParm}
\alias{initParm}
\title{Fucntion for estimating initial parameter values  'flexsurvreg'}
\usage{
initParm(CFM, DC_clean, trt)
}
\arguments{
\item{CFM}{A counter-factual model}

\item{DC_clean}{a cleaned dataset obsect obtained using dataComb.flexsurvreg}

\item{trt}{An optional additional vector denoting treatment allocations for multiple treatment comparisons.  Defaults to 'NULL'}
}
\value{
an 'optim' output giving the parameter values to be supplied as a
  starting value for the mcmc routine.
}
\description{
A generic function for estimating the initial parameters for estimation as part
of the pscfti.R function.  Parameter estimates are obtained using standard
optomisation methods provided by the 'optim' set of functions.  For a single
parameter a Brent method is applied.  For mutliple treatment comparisons the
'BFGS' method is applied.
}
\details{
This function takes the liklihood for a 'flexsurvreg' model and uses 'optim'
  to fit the likelihood.
}
