% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.flexsurvreg.R
\name{lik.flexsurvreg}
\alias{lik.flexsurvreg}
\title{Likelihood function for a psc model of class 'flexsurvreg'}
\usage{
lik.flexsurvreg(beta, pscOb)
}
\arguments{
\item{beta}{a parameter to be estimate}

\item{pscOb}{A pscOb object containing a cleaned dataset including covariates to match the CFM}
}
\value{
the results of a likelihood functions
}
\description{
A function which defines the likelihood for a PSC model where the Counter
Factual Model (CFM) takes the form of a 'flexsurvreg' object and an efficacy
parameter (\eqn{\beta}) is being estimated.  For more details on fitting please see
?pscfit and ?pscEst
}
\details{
A likelihood function for use by pscfit for a model of class
'flexsurvreg'
}
