% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{lik.fpm}
\alias{lik.fpm}
\title{Likehood functions for the estimation of psc}
\usage{
lik.fpm(beta, cov_co, haz_co, model.extract, DC.clean)
}
\arguments{
\item{beta}{efficacy parameter}

\item{cov_co}{model coefficients}

\item{haz_co}{baselined hazard coefficients}

\item{model.extract}{model extract}

\item{DC.clean}{cleaned dataset}
}
\description{
Likehood functions for the estimation of psc
}
