% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.glm.mtc.R
\name{lik.glm.mtc}
\alias{lik.glm.mtc}
\title{Likelihood functio for a a psc model of class 'flexsurvreg'}
\usage{
lik.glm.mtc(beta, DC_clean)
}
\arguments{
\item{beta}{a parameter to be estimate}

\item{DC_clean}{a cleaned dataset including covariates to match the CFM}
}
\description{
Likelihood functio for a a psc model of class 'flexsurvreg'
}
\details{
A likelihood function for use by pscfit for a model of class 'flexsurvreg'
    where mulitple treatment comparisons are
}
