% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lik.glm.R
\name{lik.glm}
\alias{lik.glm}
\title{Likelihood function for a psc model of class 'glm'}
\usage{
lik.glm(beta, pscOb)
}
\arguments{
\item{beta}{a parameter to be estimate}

\item{pscOb}{a pscOb object containing a cleaned dataset including
covariates to match the CFM}
}
\value{
the results of a likelihood functions
}
\description{
A function which defines the likelihood for a PSC model where the Counter
Factual Model (CFM) takes the form of a 'glm' object and an efficacy
parameter (\eqn{\beta}) is being estimated.  For more details on fitting please see
?pscfit and ?pscEst
}
\details{
A likelihood function for use by pscfit for a model of class
'glm'
}
