% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linPred.R
\name{linPred}
\alias{linPred}
\title{Estimates the linear predictor of a psc object}
\usage{
linPred(DC_clean, resp = F)
}
\arguments{
\item{DC_clean}{a cleaned data obhject created using dataComb()}

\item{resp}{detailing whether the linear predictor shoudl be returned on the natural or response level.  Defaults to the natural scale (resp=F)}
}
\description{
Estimates the linear predictor of a psc object
}
\details{
A function which combines the data from the data cohort against the model parameters of the PSC
}
