% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lp_psc.R
\name{lp_psc}
\alias{lp_psc}
\title{lp_psc}
\usage{
lp_psc(x)
}
\arguments{
\item{x}{an object of class 'psc'}
}
\value{
linear predictor
}
\description{
lp_psc
}
\details{
A function to estimate the linear predictor of a 'psc' object
}
