% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.extract.R
\name{model.extract.fpm}
\alias{model.extract.fpm}
\title{Model Extract}
\usage{
model.extract.fpm(CFM)
}
\arguments{
\item{x}{an R model object (of type 'flexsurvspline' or 'glm')}
}
\value{
Extracted model components
}
\description{
Model Extract
}
