% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelExtract.glm.R
\name{modelExtract.glm}
\alias{modelExtract.glm}
\title{A generic function for extracting model information}
\usage{
\method{modelExtract}{glm}(CFM)
}
\arguments{
\item{CFM}{a model of class 'glm'}
}
\value{
a list of extracted model components.  Included are
\itemize{
\item{model.frame a dataset of the covareits used to create the model}
\item{cov_co: covariate coefficients}
\item{sig: variance-covariance matrix}
\item{form: model formula}
\item{family: model family}
\item{out.nm: outcome covariates names}
}
}
\description{
A generic function for extracting model information
}
\details{
A function for extracting the model information required for using pscfit
}
