% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelExtract.R
\name{modelExtract}
\alias{modelExtract}
\title{A generic function for extracting model information}
\usage{
modelExtract(CFM)
}
\arguments{
\item{CFM}{a model object supplied to pscfit}
}
\description{
A generic function for extracting model information
}
