% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numVis.R
\name{numVis}
\alias{numVis}
\title{Visualising Categorical Data}
\usage{
numVis(x, nm)
}
\arguments{
\item{x}{a covariate to be summarised}

\item{nm}{a covariate name}
}
\value{
a ggplot object
}
\description{
A function which summarises categorical data using a waffle plot
}
