% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numVisComp.R
\name{numVisComp}
\alias{numVisComp}
\title{Visualising Numerical Data}
\usage{
numVisComp(p, x)
}
\arguments{
\item{p}{a ggplot object}

\item{x}{a numeric vector}
}
\value{
a ggplot object
}
\description{
A function which summarises numerical data using a density plots
}
