% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.psc.flexsurvreg.R
\name{plot.psc.flexsurvreg}
\alias{plot.psc.flexsurvreg}
\title{Function for Plotting PSC objects}
\usage{
\method{plot}{psc.flexsurvreg}(x, ...)
}
\arguments{
\item{x}{an object of class 'psc'}

\item{...}{not used}
}
\value{
a survival plot corresponding to the psc fit
}
\description{
Function for Plotting PSC objects
}
\details{
making use of 'ggsurvplot' in the survminer package, this function
plots the expected survival funtion for the 'control' treatment estimated
from the CFM along with the Kaplan Meier estimates of the observed events
}
