% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ite.R
\name{plot_ite}
\alias{plot_ite}
\title{Function for Plotting individual treatment effects for a PSC object}
\usage{
plot_ite(x)
}
\arguments{
\item{x}{an object of class 'psc'}
}
\value{
a plot corresponding to the psc fit
}
\description{
Function for Plotting individual treatment effects for a PSC object
}
\details{
a funtion which returns a plot showing the individual estimated
response under the control treatment (from the CFM) against the observed
response from the data cohort
}
