% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.psc}
\alias{print.psc}
\title{Personalised Synthetic Controls - print}
\usage{
\method{print}{psc}(x, ...)
}
\arguments{
\item{x}{an object of class 'psc'}

\item{...}{not used}
}
\value{
printing psc results
}
\description{
Personalised Synthetic Controls - print
}
