% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.quiet_list.R
\name{print.quiet_list}
\alias{print.quiet_list}
\title{quiet_gtsumm}
\usage{
\method{print}{quiet_list}(x, ...)
}
\arguments{
\item{x}{an object of class 'psc'}

\item{...}{not used}
}
\value{
A quiet list
}
\description{
Ensuring a quiet list of the grobs data are supplied to cfmDataVis
}
