% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc-package.R
\docType{package}
\name{psc-package}
\alias{psc}
\alias{psc-package}
\title{psc: Personalised Synthetic Controls}
\description{
Allows the comparison of data cohorts (DC) against a Counter Factual Model (CFM) and measures the difference in terms of an efficacy parameter. Allows the application of Personalised Synthetic Controls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/richjjackson/psc/}
  \item \url{https://github.com/richJJackson/psc}
  \item Report bugs at \url{https://github.com/richJJackson/psc/issues}
}

}
\author{
\strong{Maintainer}: Richard Jackson \email{richJ23@liverpool.ac.uk} (\href{https://orcid.org/0000-0002-7814-5088}{ORCID}) [copyright holder]

}
\keyword{internal}
