% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc.object.R
\docType{data}
\name{psc.object}
\alias{psc.object}
\title{Fitted \code{psc} object}
\format{
An object of class \code{NULL} of length 0.
}
\usage{
psc.object
}
\description{
An object returned by the \code{pscfit} function, inheriting from class
\code{psc} and representing a fitted personalised synthetic control model.
}
\author{
Richard Jasckson (\email{richj23@liverpool.ac.uk})
}
\keyword{datasets}
