% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psc_ite.R
\name{psc_ite}
\alias{psc_ite}
\title{Wrapper function for individual treatment effects}
\usage{
psc_ite(model, data)
}
\arguments{
\item{model}{a CFM}

\item{data}{a data cohort}
}
\description{
Wrapper function for individual treatment effects
}
\details{
A wrapper function that estimates individual treatment effects usinf pscfit
}
