% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscCFM.R
\name{pscCFM}
\alias{pscCFM}
\title{A function which provides basic summary information of a matrix x}
\usage{
pscCFM(cfm, covnm = NULL, setting = NULL, valid = NULL, citation = NULL)
}
\arguments{
\item{cfm}{a model object}

\item{covnm}{an optinal character vector suppling covariate names to the terms
included in the model}

\item{setting}{and object from 'cfmSett' detailing the setting of the model.}

\item{valid}{an object from cfmValid which details validation of the model}

\item{citation}{a character giving a reference (if applicable),  defaults to NULL (e.g no citation)}
}
\description{
A function which provides basic summary information of a matrix x
}
\details{
Categorical/Character data are summarised by a table and Continuous
    data are summarised as median (IQR)
}
