% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscComb.R
\name{pscComb}
\alias{pscComb}
\title{PSC Estiamtion combination}
\usage{
pscComb(x)
}
\arguments{
\item{x}{A pscfit object with 2 treatment comparions}
}
\description{
This function allows the combination of estimation where mulitple treatment
comparisons are being made and one of the treatment levels relates to the
control level.  Note here it is assumed that the control level is coded
trt=0. This function then combines the two efficacy estimates for the
treatment effect (concurrent and synhtetic) whilst accounting for the
correlation inherent in these two estimates.
}
