% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscData_addLik.R
\name{pscData_addLik}
\alias{pscData_addLik}
\title{A function that add a likelihood for estimation to the pscObject}
\usage{
pscData_addLik(CFM)
}
\arguments{
\item{CFM}{A counter factual model}
}
\value{
a likelihood function
}
\description{
The purpose of this function is to include the appropriate likelihood to the
psc object for estimation procedures
}
