% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscData_addtrt.R
\name{pscData_addtrt}
\alias{pscData_addtrt}
\title{A function that includes a treatment indicator when multiple
treatment comparisons are required}
\usage{
pscData_addtrt(DC, trt)
}
\arguments{
\item{DC}{a data cohort to be 'cleaned'}

\item{trt}{a treatment indicator}
}
\value{
a dataset which is checked and compatible with the CFM
}
\description{
The purpose of this function is to organise a treatment indicator where
multiple treatment comparisons are being evaluated. This acts as a
sub-function to the pscData.R function.
}
