% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscData_error.R
\name{pscData_error}
\alias{pscData_error}
\title{A function which performs error checks between the DC and CFM}
\usage{
pscData_error(term.nm, DC)
}
\arguments{
\item{term.nm}{Term names from the CFM}

\item{DC}{a data cohort to be 'cleaned'}
}
\value{
a 'stop' command when errors are detected
}
\description{
The purpose of this function is check that terms included in the Data Cohort
match those used within the Counter Factual Model. This acts as a
sub-function to the pscData.R function.
}
