% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscData_match.R
\name{pscData_match}
\alias{pscData_match}
\title{A function to ensure that data from the cfm and data cohort are compatible}
\usage{
pscData_match(cls, lev, DC)
}
\arguments{
\item{cls}{a list of extracted data classes}

\item{lev}{a list of factor levels}

\item{DC}{a data cohort to be 'cleaned'}
}
\value{
a dataset which is checked and compatible with the CFM
}
\description{
The purpose of this function is to run a series of checks to ensure that the
data included in the data cohort is comparable to the counter-factual model.
This matches the data classes and checks the levels in the DC match those
used in the CFM. This acts as a sub-function to the pscData.R function.
}
