% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscData_miss.R
\name{pscData_miss}
\alias{pscData_miss}
\title{A function which removes missing data from the DC}
\usage{
pscData_miss(DC)
}
\arguments{
\item{DC}{a data cohort to be 'cleaned'}
}
\value{
a dataset with missing data removed
}
\description{
Currently the psc package works only on complete-case datasets.  This
function removes rows with missing data and returns a warning to inform the
user. This acts as a sub-function to the pscData.R function.
}
