% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscData_structure.R
\name{pscData_structure}
\alias{pscData_structure}
\title{A function which structures the Data Cohort in a format for model estimation}
\usage{
pscData_structure(CFM, DC)
}
\arguments{
\item{CFM}{a Counter Factual Model}

\item{DC}{a Data Cohort object}
}
\value{
A set of structures for use with estimation procedures
re-export Surv from survival
}
\description{
This function ensures the data are supplied in a structure which allows for
estimation.  This is performed by re-fitting the original CFM with the DC and
extracting the appropriate structures.  Data are returned in terms of "Y" for
model outcomes, "X" for data and "Z" for random effects where mixed models
are supplied.
}
