% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscEst.pscCFM.R
\name{pscEst.pscCFM}
\alias{pscEst.pscCFM}
\title{Function for estimating initial parameter values  'flexsurvreg'}
\usage{
\method{pscEst}{pscCFM}(CFM, DC_clean, nsim, start, start.se, trt = NULL)
}
\arguments{
\item{CFM}{a model object supplied to pscfit}

\item{DC_clean}{a cleaned dataset ontained using dataComb().}

\item{nsim}{the number of MCMC simulations to run}

\item{start}{the stating value for}

\item{start.se}{the stating value fo}

\item{trt}{an optional vector denoting treatment allocations where multiple
treatment comparisons are being made}
}
\value{
A matrix containing the draws form the posterior distribution
}
\description{
A function which performs the Bayesian MCMC estimation procedure for
estimating the efficacy parameter (\eqn{\beta}) using personalised sunthetic
controls methodology.
}
\details{
An MCMC routine for fitting a psc model
}
