% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscEst_samp.R
\name{pscEst_samp}
\alias{pscEst_samp}
\title{Starting conditions for Bayesian MCMC estimation procedures in 'pscfit'
A procedure which runs the sampling process for MCMC estimation}
\usage{
pscEst_samp(pscOb, nsim)
}
\arguments{
\item{pscOb}{an pscOb object which has been passed through pscData() and
init() functions}

\item{nsim}{the number of MCMC simulations to run}
}
\value{
An updated set of attributes for the pscOb which includes
}
\description{
Starting conditions for Bayesian MCMC estimation procedures in 'pscfit'
A procedure which runs the sampling process for MCMC estimation
}
