% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscEst_update.R
\name{pscEst_update}
\alias{pscEst_update}
\title{Updating the posterior distribution as part of the MCMC estimation process
A procedure which performs a single update of the posterior distribution}
\usage{
pscEst_update(i, draws, pscOb)
}
\arguments{
\item{i}{index of the draw number (i>1)}

\item{draws}{a matrix containing the posterior draws to update}

\item{pscOb}{an pscOb object which has been passed through pscData() and
init() functions}
}
\value{
An updated set of posterior draws
}
\description{
Updating the posterior distribution as part of the MCMC estimation process
A procedure which performs a single update of the posterior distribution
}
