% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscfitsub.R
\name{pscfitsub}
\alias{pscfitsub}
\title{Wrapper function for sub-group analysis.}
\usage{
pscfitsub(CFM, DC, sg, cont_lev = 2, cont_cut = NULL)
}
\arguments{
\item{CFM}{CFM}

\item{DC}{DC}

\item{sg}{sg}

\item{cont_lev}{cont_lev}

\item{cont_cut}{cont_cut}
}
\description{
Wrapper function for sub-group analysis.
}
