% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscsub.R
\name{pscsub}
\alias{pscsub}
\title{sub group analyses using PSC}
\usage{
pscsub(CFM, DC, sg, cont_lev = 2, cont_cut = NULL)
}
\arguments{
\item{CFM}{CFM}

\item{DC}{DC}

\item{sg}{sg}

\item{cont_lev}{cont_lev}

\item{cont_cut}{cont_cut}
}
\description{
sub group analyses using PSC
}
