% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscsubWrap.R
\name{pscsubWrap}
\alias{pscsubWrap}
\title{Wrapper for sub-group analysis}
\usage{
pscsubWrap(pscsub)
}
\arguments{
\item{CFM}{CFM}

\item{DC}{DC}

\item{sg}{sg}

\item{cont_lev}{cont_lev}

\item{cont_cut}{cont_cut}
}
\description{
Wrapper for sub-group analysis
}
