% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscSumm.R
\name{pscSumm}
\alias{pscSumm}
\title{Personalised Synthetic Controls - summary}
\usage{
pscSumm(DC_clean)
}
\arguments{
\item{DC_clean}{a cleaned dataset ontained using dataComb().}
}
\value{
psc summary results including an estimate of the linear predictor
combing the data and the model, an estimate of patient level response and
summary statistics of the average responses for the sythenthic and observed
populations
}
\description{
A function which provides a summary of a DC_clean object.
To be used either in conjunction with dataComb.R or summary.psc.R
}
\examples{
library(psc)
library(survival)
data("surv.mod")
data("data")
dc <- dataComb(surv.mod,data)
pscSumm(dc)
}
