% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_surv_est.R
\name{spline_surv_est}
\alias{spline_surv_est}
\title{Counter Factual Model - summary}
\usage{
spline_surv_est(lam, kn, k, haz_co, cov_co, cov = cov, tm = tm, beta = 0)
}
\arguments{
\item{lam}{parameters of the flexible spline model}

\item{kn}{knots included in the flexible spline model}

\item{k}{number of knots in the flexible spline model}

\item{haz_co}{parameters for the baseline hazard function in the flexible
spline model}

\item{cov_co}{covariate parameters of the flexible spline model}

\item{cov}{a matrix of covaraites from the Data Cohort}

\item{tm}{time at which to assess the survival function}

\item{beta}{parameter with which to adjust the baseline function (defaults to
beta=0)}
}
\value{
A data frame containing survival estimates for a give time
}
\description{
A function to estimate the survival function based on parameter estimates -
used in ootstrapping CFM for CIs
}
