% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visComp.R
\name{visComp}
\alias{visComp}
\title{Visualising Comparisons between a CFM and a DC}
\usage{
visComp(CFM, DC, id = NULL)
}
\arguments{
\item{CFM}{an object of class pscCFM}

\item{DC}{A dataset including columns to match to covariates in the model}

\item{id}{Numeric vector stating which patient(s) from the dataset should be included in the analysis.
Defaults to all patients}
}
\value{
a list of grobs for each model covariate
}
\description{
The visComp function takes the data visualisations supplied as part of the
CFM model and appends summaries of the equivalent datapoints from the
Data Cohort.
}
