% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_lp.R
\name{boot_lp}
\alias{boot_lp}
\title{Counter Factual Model - summary}
\usage{
boot_lp(i, pscOb, resp = resp)
}
\arguments{
\item{pscOb}{an object of class 'psc'}

\item{resp}{A boolean object to determine if results should be presented on
the response scale}
}
\value{
A summary of a cfm object
}
\description{
A function to estimate the linear predictor - used in bootstrapping CFM for
CIs
}
