% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boot_sest.R
\name{boot_sest}
\alias{boot_sest}
\title{Counter Factual Model - summary}
\usage{
boot_sest(
  i,
  pscOb = pscOb,
  lam = lam,
  kn = kn,
  k = k,
  cov = cov,
  tm = tm,
  rest = rest,
  beta = beta
)
}
\arguments{
\item{object}{an object of class 'psc'}
}
\value{
A summary of a cfm object
}
\description{
A function to estimate the survival function based on parameter estimates -
used in ootstrapping CFM for CIs
}
