% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmDataSumm.R
\name{cfmDataSumm}
\alias{cfmDataSumm}
\title{Summarising data within a Counter Factual Model (CFM)}
\usage{
cfmDataSumm(cfm)
}
\arguments{
\item{cfm}{a 'glm' or 'flexsurvreg' model object}
}
\value{
a summary table
}
\description{
The pscCFM creates a model object which is stripped of identifiable
information.  The cfmDataSumm function supplies a tabulated form of the
dataset used in the CFM for summary information.  Information returned in
the form of a table
}
