% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfmSumm.glm.R
\name{cfmSumm.glm}
\alias{cfmSumm.glm}
\title{Counter Factual Model - summary}
\usage{
cfmSumm.glm(pscOb, bootCI = TRUE, nboot = 1000, resp = TRUE)
}
\arguments{
\item{pscOb}{an object of class 'psc'}

\item{bootCI}{a boolean to determine if bootstrapping CIs are required}

\item{nboot}{Number of bootstraps}

\item{resp}{Should results be on the response scale?}
}
\value{
A summary of a cfm object
}
\description{
A generic function to provide a summary of a Counter factual model of class
'glm'
}
