% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.psc.R
\name{coef.psc}
\alias{coef.psc}
\title{Returns the coefficient estimate of a psc object.}
\usage{
\method{coef}{psc}(object, ...)
}
\arguments{
\item{object}{a 'psc' object}

\item{...}{not used}
}
\value{
The summary of the posterior distribution for the efficacy parameter
in terms of the median and 95% HPD
}
\description{
Returns basic measures of the posterior distribution obtained from the
psc object
}
