% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.psc.R
\name{coef.psc}
\alias{coef.psc}
\title{Returns the coefficeint estimate of a psc object.}
\usage{
\method{coef}{psc}(object, ..., level = 0.05)
}
\arguments{
\item{object}{a 'psc' object}

\item{...}{not used}

\item{level}{the level at which credibility intervals are assessed, defaults to 0.05}
}
\value{
The summary of the posterior distribution for the efficacy parameter in terms of the median and 95% HPD
}
\description{
Returns basic measures of the posterior distribution obtained from the
pscfit.R function
}
