% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_match.R
\name{data_match}
\alias{data_match}
\title{A function to ensure that data from the cfm and data cohort are compatible}
\usage{
data_match(cfm.data, dc.data)
}
\arguments{
\item{cfm.data}{a model object supplied to pscfit}

\item{dc.data}{a dataset including covariates to match the CFM}
}
\value{
a dataset which is checked and compatible with the CFM
}
\description{
The purpose of this function is to run a series of checks to ensure that the
data included in the data cohort is comparable to the counter-factual model
}
