% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataComb.R
\name{dataComb}
\alias{dataComb}
\title{A generic function for cleaning data ready for analysis}
\usage{
dataComb(CFM, DC, id = NULL, trt = NULL)
}
\arguments{
\item{CFM}{a model object supplied to pscfit}

\item{DC}{a dataset including covariates to match the CFM}

\item{id}{to specify which observations in the data cohort should be evaluated.  Defualts to 'NULL' i.e all observations}

\item{trt}{used to specify multiple treatment effects. Defaults to NULL}
}
\description{
A generic function for cleaning data ready for analysis
}
\examples{
library(psc)
library(survival)
data("surv.mod")
data("data")
dc <- dataComb(surv.mod,data)
}
