% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataComb.R
\name{dataComb}
\alias{dataComb}
\title{A generic function for cleaning data ready for analysis}
\usage{
dataComb(CFM, DC, id = NULL, trt = NULL)
}
\arguments{
\item{CFM}{a model object supplied to pscfit}

\item{DC}{a dataset including covariates to match the CFM}
}
\description{
A generic function for cleaning data ready for analysis
}
