% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init}
\alias{init}
\title{Function for estimating initial parameter values}
\usage{
init(pscOb)
}
\arguments{
\item{pscOb}{a psc object}
}
\value{
Parameter Estimates and standard error for the efficacy parameter
}
\description{
Function for estimating initial parameter values
}
\details{
This function takes the likelihood and data structures provided by the
pscData() strucutres and fits the likelihood to provide starting values for
MCMC estimation
}
\examples{
e4_data <- psc::e4_data
gemCFM <- psc::gemCFM
pscOb <- pscData(gemCFM,e4_data)
pscOb <- init(pscOb)
}
