% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initParm.R
\name{initParm}
\alias{initParm}
\title{Fucntion for estimating initial parameter values  'flexsurvreg'}
\usage{
initParm(CFM, DC_clean, trt)
}
\arguments{
\item{CFM}{A counter-factual model}

\item{DC_clean}{a cleaned dataset obsect obtained using dataComb.flexsurvreg}

\item{trt}{An optional additional vector denoting treatment allocations for multiple treatment comparisons.  Defaults to 'NULL'}
}
\value{
an 'optim' output giving the parameter values to be supplied as a
  starting value for the mcmc routine.
}
\description{
Fucntion for estimating initial parameter values  'flexsurvreg'
}
\details{
This function takes the liklihood for a 'flexsurvreg' model and uses 'optim'
  to fit the likelihood.
}
