% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelExtract.flexsurvreg.R
\name{modelExtract.flexsurvreg}
\alias{modelExtract.flexsurvreg}
\title{A generic function for extracting model information}
\usage{
\method{modelExtract}{flexsurvreg}(CFM)
}
\arguments{
\item{CFM}{a model of class either 'flexsurvreg'}
}
\value{
a list of extracted model components
}
\description{
A generic function for extracting model information
}
\details{
A function for extracting the model information required for using pscfit
}
