% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelExtract.R
\name{modelExtract}
\alias{modelExtract}
\title{A generic function for extracting model information}
\usage{
modelExtract(CFM)
}
\arguments{
\item{CFM}{a model of class either 'glm' or 'flexsurvreg'}
}
\value{
a list of extracted model components
}
\description{
A generic function for extracting model information
}
\details{
A function for extracting the model information required for using pscfit
}
