% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modp.R
\name{modp}
\alias{modp}
\title{modp}
\usage{
modp(x)
}
\arguments{
\item{x}{a numberic vector}
}
\value{
a numeric vector with negative values replaced with 0
}
\description{
A function which rrturns either the input value (if positive) or zero (if
negative)
}
\details{
A fucntion which returns a version of x with negative values replacd with 0
}
