% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.psc.binary.R
\name{plot.psc.binary}
\alias{plot.psc.binary}
\title{Function for Plotting PSC objects}
\usage{
\method{plot}{psc.binary}(x, ...)
}
\arguments{
\item{x}{an object of class 'psc'}

\item{...}{not used}
}
\value{
a survival plot corresponding to the psc fit
}
\description{
A function which illsutrates the predicted response under the counter factual
model and the observed response under the experimental treatment(s).
}
\details{
This function plots the expected response of the control treatment
   along with the observe response rates of the experimental arms
}
\examples{
bin.mod <- psc::bin.mod
data <- psc::data
bin.psc <- pscfit(bin.mod,data,nsim=3000)
plot(bin.psc)
}
