% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.psc.binary.R
\name{plot.psc.binary}
\alias{plot.psc.binary}
\title{Function for Plotting PSC objects}
\usage{
\method{plot}{psc.binary}(x)
}
\arguments{
\item{x}{an object of class 'psc'}
}
\value{
a survival plot corresponding to the psc fit
}
\description{
Function for Plotting PSC objects
}
\details{
This function plots the expected response of the control treatment
   along with the observe response rates of the experimental arms
}
