% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.psc}
\alias{plot.psc}
\title{Fucntion for Plotting PSC objects}
\usage{
\method{plot}{psc}(x, ...)
}
\arguments{
\item{x}{an object of class 'psc'}
}
\value{
a plot corresponding to the psc fit
}
\description{
Fucntion for Plotting PSC objects
}
\examples{
psc.ob <- psc(model,data)
plot(psc.ob)
}
