% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.psc}
\alias{plot.psc}
\title{Function for Plotting PSC objects}
\usage{
\method{plot}{psc}(x)
}
\arguments{
\item{x}{an object of class 'psc'}
}
\value{
a plot corresponding to the psc fit
}
\description{
Function for Plotting PSC objects
}
\details{
making use of the generic 'plot' functions this will provide some
  graphical output of the fitted psc object.  The form of the output will
  depend on the class of the initial model
}
