% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ite.glm.R
\name{plot_ite.glm}
\alias{plot_ite.glm}
\title{Function for Plotting PSC objects}
\usage{
plot_ite.glm(x)
}
\arguments{
\item{x}{an object of class 'psc'}
}
\value{
A plot showing the individual treatment effects
}
\description{
Function for Plotting PSC objects
}
\details{
This function plots the expected response of the control treatment
   along with the observe response rates for each patient in the dataset
}
