% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ite.R
\name{plot_ite}
\alias{plot_ite}
\title{Function for Plotting individual treatment effects for a PSC object}
\usage{
plot_ite(x)
}
\arguments{
\item{x}{an object of class 'psc'}
}
\value{
A plot showing the individual treatment effects
}
\description{
A function which plots a patients response against their predicted response
from a CFM.  Exact form of the output will depend on the form of the model supplied
}
\details{
This function plots the expected response of the control treatment
   along with the observe response rates for each patient in the dataset
}
\examples{
library(psc)
library(survival)
bin.mod <- psc::bin.mod
data <- psc::data
bin.psc <- pscfit(bin.mod,data)
plot_ite(bin.psc)
}
