% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscCFM.R
\name{pscCFM}
\alias{pscCFM}
\title{Creating a CFM model which can be shared}
\usage{
pscCFM(CFM, dataSumm = T, dataVis = T)
}
\arguments{
\item{CFM}{a 'glm' or 'flexsurvreg' model object}

\item{dataSumm}{a logical indicator specifying whether a summary of the data
should be provided, defaults to TRUE.}

\item{dataVis}{a logical indicator specifying whether a visualisations of the
data should be provided, defaults to TRUE.}
}
\value{
a list containing objects which specify the required exported components
  of the model.
}
\description{
Standard R model objects contain within them the datasets used to create the
model and as such care is needed when sharing these objects for research.
The psc.cfm function creates an object with all identifiable information
retracted and includes only the information required to use the models within
the psc package
}
