% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pscfit.R
\name{pscfit}
\alias{pscfit}
\title{Personalised Synthetic Controls}
\usage{
pscfit(CFM, DC, nsim = 5000)
}
\arguments{
\item{nsim}{The number of simulations for the MCMC routine}

\item{model}{A model of type 'glm' or 'flexsurvspline'}

\item{data}{A dataset including columns to match to covariates in the model}
}
\value{
a object of class 'psc'
}
\description{
Personalised Synthetic Controls
}
\examples{
psc.ob <- psc(model,data)
summary(psc.ob)
}
