% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.psc}
\alias{summary.psc}
\title{Personalised Synthetic Controls - summary}
\usage{
\method{summary}{psc}(x, ...)
}
\arguments{
\item{x}{an object of class 'psc'}
}
\value{
psc summary results'
}
\description{
Personalised Synthetic Controls - summary
}
