% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.psc}
\alias{summary.psc}
\title{Personalised Synthetic Controls - summary}
\usage{
\method{summary}{psc}(object, ...)
}
\arguments{
\item{object}{an object of class 'psc'}

\item{...}{not used}
}
\value{
A summary of a psc object obtained using pscSumm and a copy of the pscfit object
}
\description{
A generic function to provide a summary of a 'psc' object obtained from
pscfit.R
}
\examples{
library(psc)
library(survival)
data("surv.mod")
data("data")
psc.ob <- pscfit(surv.mod,data)
summary(psc.ob)
}
